Add files via upload
had to do a weird upload because apparently you're supposed to only share the src through git (not the other folder because it's like filepaths and stuff), so my clone and project were not right, so colin had to make a new copy somewhere else on my drive to get it to work but now my version doesn't have the git functionality lol. hopefully this upload doesn't totally fuck everything?

anyway, cron.xml is right. CronServlet.java needs to change a lot (just started copying BlogServlet but haven't changed much yet), and I'm not sure of how to do the servlet mapping in web.xml. might have something to do with using the url path I put in cron.xml?